Major simplifications and some performance improvements.

+++ src/main/java/myworld/hummingbird/DebugHandler.java
@@ -2,6 +2,6 @@ package myworld.hummingbird;
 
 public interface DebugHandler {
 
-    void debug(HummingbirdVM vm, Fiber fiber, int staticValue, int dynamicValue);
+    void debug(HummingbirdVM vm, Fiber fiber, int staticValue, long dynamicValue);
 
 }
